IngredientService changed to use repository

+++ src/main/java/es/ssdd/PracticaSSDD/service/IngredientService.java
@@ -51,6 +51,7 @@ public class IngredientService {
     }
 
     public boolean removeIngredient(Long id) {
+        if(!ingredientRepository.existsById(id)) return false;
         ingredientRepository.deleteById(id);
         return true;
     }
